Conflicts:
	AUTHORS
	README.md
	src/core/ArxGame.cpp

+++ src/animation/Skeleton.h
@@ -29,8 +29,8 @@
 
 struct VertexGroup {
 	std::string       name;
-	long              origin;
-	std::vector<long> indexes;
+	size_t            origin;
+	std::vector<size_t> indexes;
 	float             siz;
 	
 	VertexGroup()
@@ -48,7 +48,7 @@ struct BoneTransform {
 };
 
 struct Bone {
-	std::vector<long> idxvertices;
+	std::vector<size_t> idxvertices;
 	long              father;
 	
 	BoneTransform     anim;
